Move GroupByKey expansion into DirectPipelineRunner
